+++ js/locale/dom-locale-loader.js
@@ -5,6 +5,10 @@ export class DomLocaleLoaderPlugin extends RipeCommonsPlugin {
         await super.load();
     }
 
+    getVersion() {
+        return "0.1.0";
+    }
+
     getCapabilities() {
         return [RipeCommonsCapability.new("locale-loader")];
     }