+++ js/util/device.js
@@ -10,6 +10,10 @@ export class DevicePlugin extends RipeCommonsPlugin {
         await super.unload();
     }
 
+    getVersion() {
+        return "0.1.0";
+    }
+
     getCapabilities() {
         return [RipeCommonsCapability.new("helper")];
     }