+++ js/util/url-changer.js
@@ -27,6 +27,10 @@ export class UrlChangerPlugin extends RipeCommonsPlugin {
         await super.unload();
     }
 
+    getVersion() {
+        return "0.1.0";
+    }
+
     getCapabilities() {
         return [RipeCommonsCapability.new("helper")];
     }